+++ src/bundles/Elsa.WorkflowServer.Web/Program.cs
@@ -16,7 +16,6 @@ using Elsa.MongoDb.Extensions;
 using Elsa.MongoDb.Modules.Identity;
 using Elsa.MongoDb.Modules.Management;
 using Elsa.MongoDb.Modules.Runtime;
-using Elsa.WorkflowServer.Web.WorkflowContexts;
 using Microsoft.Data.Sqlite;
 using Microsoft.Extensions.Options;
 using Proto.Persistence.Sqlite;
@@ -216,8 +215,6 @@ services
 #endif
     });
 
-services.AddWorkflowContextProvider<CustomerWorkflowContextProvider>();
-services.AddWorkflowContextProvider<OrderWorkflowContextProvider>();
 services.AddHealthChecks();
 services.AddControllers();
 services.AddCors(cors => cors.AddDefaultPolicy(policy => policy.AllowAnyHeader().AllowAnyMethod().AllowAnyOrigin().WithExposedHeaders("*")));
