+++ src/modules/Elsa.Workflows.Core/Models/Input.cs
@@ -15,7 +15,7 @@ public abstract class Input : Argument
     {
         Type = type;
     }
-    
+
     /// <inheritdoc />
     protected Input(Expression? expression, MemoryBlockReference memoryBlockReference, Type type) : base(memoryBlockReference)
     {
@@ -31,7 +31,8 @@ public abstract class Input : Argument
     /// <summary>
     /// Gets the type of the input.
     /// </summary>
-    [JsonPropertyName("typeName")] public Type Type { get; set; }
+    [JsonPropertyName("typeName")]
+    public Type Type { get; set; }
 }
 
 /// <summary>
@@ -43,7 +44,7 @@ public class Input<T> : Input
     public Input(MemoryBlockReference memoryBlockReference) : base(memoryBlockReference, typeof(T))
     {
     }
-    
+
     /// <inheritdoc />
     public Input(T literal, string? id = default) : this(new Literal<T>(literal, id))
     {
@@ -88,7 +89,7 @@ public class Input<T> : Input
     public Input(Literal literal) : base(Expression.LiteralExpression(literal.Value), literal, typeof(T))
     {
     }
-    
+
     /// <inheritdoc />
     public Input(ObjectLiteral<T> literal) : base(Expression.LiteralExpression(literal.Value), literal, typeof(T))
     {
@@ -101,6 +102,11 @@ public class Input<T> : Input
 
     /// <inheritdoc />
     public Input(Expression expression, MemoryBlockReference memoryBlockReference) : base(expression, memoryBlockReference, typeof(T))
+    {
+    }
+
+    /// <inheritdoc />
+    public Input(Expression expression) : this(expression, new MemoryBlockReference())
     {
     }
 }
