+++ test/integration/Elsa.IntegrationTests/Scenarios/JsonObjectToObjectRemainsJsonObject/Activities/ObjectProcessor.cs
@@ -1,0 +1,16 @@
+using Elsa.Extensions;
+using Elsa.Workflows.Core;
+using Elsa.Workflows.Core.Models;
+
+namespace Elsa.IntegrationTests.Scenarios.JsonObjectToObjectRemainsJsonObject.Activities;
+
+public class ObjectProcessor : CodeActivity<object>
+{
+    public Input<object> Value { get; set; } = default!;
+
+    protected override void Execute(ActivityExecutionContext context)
+    {
+        var value = Value.Get(context);
+        context.SetResult(value);
+    }
+}
