+++ test/integration/Elsa.IntegrationTests/Scenarios/JsonObjectToObjectRemainsJsonObject/Workflows/TestWorkflow.cs
@@ -1,11 +1,12 @@
 using System.Text.Json.Nodes;
-using Elsa.IntegrationTests.Scenarios.JsonObjectJint.Activities;
+using Elsa.Expressions.Models;
+using Elsa.IntegrationTests.Scenarios.JsonObjectToObjectRemainsJsonObject.Activities;
 using Elsa.JavaScript.Models;
 using Elsa.Workflows.Core;
 using Elsa.Workflows.Core.Activities;
 using Elsa.Workflows.Core.Contracts;
 
-namespace Elsa.IntegrationTests.Scenarios.JsonObjectJint.Workflows;
+namespace Elsa.IntegrationTests.Scenarios.JsonObjectToObjectRemainsJsonObject.Workflows;
 
 public class TestWorkflow : WorkflowBase
 {
@@ -19,6 +20,11 @@ public class TestWorkflow : WorkflowBase
             {
                 new JsonObjectProducer
                 {
+                    Result = new(item)
+                },
+                new ObjectProcessor
+                {
+                    Value = new(JavaScriptExpression.Create("getItem()"), new MemoryBlockReference()),
                     Result = new(item)
                 },
                 new WriteLine(JavaScriptExpression.Create("getItem()['Foo']['Bar']"))
